requirements.txt
selenium==4.18.1
requests==2.31.0
beautifulsoup4==4.12.3
python-nmap==0.7.1
sqlmap==1.7.7
python-owasp-zap-v2.4==0.0.14
reportlab==4.1.0
pandas==2.2.1
python-dotenv==1.0.1
colorama==0.4.6
tqdm==4.66.2 